+++ packages/sourcemap/src/require_sourcemap.js
@@ -4,6 +4,6 @@ const require = createRequire(import.meta.url)
 // consider using https://github.com/7rulnik/source-map-js
 
 export const requireSourcemap = () => {
-  const namespace = require("source-map")
+  const namespace = require("source-map-js")
   return namespace
 }