+++ tests/dev_server/errors/generate_snapshot_files.mjs
@@ -1,20 +1,23 @@
 import { ensureEmptyDirectory, writeFileSync } from "@jsenv/filesystem"
 import { chromium, firefox, webkit } from "playwright"
+import { createTaskLog } from "@jsenv/log"
 
 process.env.GENERATING_SNAPSHOTS = "true"
 const { devServer } = await import("./start_dev_server.mjs")
 const snapshotDirectoryUrl = new URL(`./snapshots/`, import.meta.url)
 const screenshotsDirectoryUrl = new URL(`./sceenshots/`, import.meta.url)
-await ensureEmptyDirectory(snapshotDirectoryUrl)
+if (!process.env.FROM_TESTS) {
+  await ensureEmptyDirectory(snapshotDirectoryUrl)
+}
 await ensureEmptyDirectory(screenshotsDirectoryUrl)
 
 const test = async ({ browserLauncher, browserName }) => {
   const browser = await browserLauncher.launch({ headless: true })
 
-  const generateHtmlForStory = async ({
-    story,
-    preferServerErrorReporting,
-  }) => {
+  const generateHtmlForStory = async ({ story }) => {
+    const task = createTaskLog(`snapshoting ${story} on ${browserName}`, {
+      disabled: process.env.FROM_TESTS,
+    })
     const page = await browser.newPage()
     await page.goto(`${devServer.origin}/${story}/main.html`)
     try {
@@ -24,10 +27,9 @@ const test = async ({ browserLauncher, browserName }) => {
         `jsenv error overlay not displayed on ${browserName} for ${story}`,
       )
     }
-    if (preferServerErrorReporting) {
-      // wait a bit more to let server error replace browser error
-      await new Promise((resolve) => setTimeout(resolve, 200))
-    }
+    // wait a bit more to let client time to fetch error details from server
+    await new Promise((resolve) => setTimeout(resolve, 200))
+
     const htmlGenerated = await page.evaluate(
       /* eslint-disable no-undef */
       async () => {
@@ -43,14 +45,16 @@ const test = async ({ browserLauncher, browserName }) => {
       },
       /* eslint-enable no-undef */
     )
-    await page.setViewportSize({ width: 900, height: 550 }) // generate smaller screenshots
-    const sceenshotBuffer = await page
-      .locator("jsenv-error-overlay")
-      .screenshot()
-    writeFileSync(
-      new URL(`./${story}_${browserName}.png`, screenshotsDirectoryUrl),
-      sceenshotBuffer,
-    )
+    if (!process.env.FROM_TESTS) {
+      await page.setViewportSize({ width: 900, height: 550 }) // generate smaller screenshots
+      const sceenshotBuffer = await page
+        .locator("jsenv-error-overlay")
+        .screenshot()
+      writeFileSync(
+        new URL(`./${story}_${browserName}.png`, screenshotsDirectoryUrl),
+        sceenshotBuffer,
+      )
+    }
     writeFileSync(
       new URL(`./${story}_${browserName}.html`, snapshotDirectoryUrl),
       process.platform === "win32"
@@ -58,6 +62,7 @@ const test = async ({ browserLauncher, browserName }) => {
         : htmlGenerated,
     )
     await page.close()
+    task.done()
   }
 
   try {
@@ -66,7 +71,6 @@ const test = async ({ browserLauncher, browserName }) => {
     })
     await generateHtmlForStory({
       story: "js_import_not_found",
-      preferServerErrorReporting: true,
     })
     await generateHtmlForStory({
       story: "js_import_syntax_error",
@@ -76,14 +80,12 @@ const test = async ({ browserLauncher, browserName }) => {
     })
     await generateHtmlForStory({
       story: "plugin_error_transform",
-      preferServerErrorReporting: true,
     })
     await generateHtmlForStory({
       story: "script_module_inline_export_not_found",
     })
     await generateHtmlForStory({
       story: "script_module_inline_import_not_found",
-      preferServerErrorReporting: true,
     })
     await generateHtmlForStory({
       story: "script_module_inline_syntax_error",