+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 27.6.1
+
+- code frame show source code frame also in firefox/webkit for runtime errors
+- use [sourcemap-js](https://github.com/7rulnik/source-map-js) instead of [source-map](https://github.com/mozilla/source-map)
+
 # 27.6.0
 
 - add code frame in error overlay for fetch/parse/execute errors